feat(cli): retire the proxy chat engine, gate dev commands out of release builds
Change 3 of doc 14 — the sequence completes: one chat (the TUI), clean tree.

- modules/intelligence/ deleted (the engine's only importer was the boot-
  helpers line; zero consumers of its bus events): proxy model helpers live
  in modules/proxy/models.ts, the sessions command in modules/analysis/
  (sessions are live launch-identity rows; messages/parts history is frozen
  legacy data)
- the BusEvent union shrinks to the provenance members; dead summarizer
  branches removed
- dev command umbrella: chat/profile/run register only in the dev channel —
  INFLEXA_BUILD_CHANNEL joins bakedEnv (release builds must declare it or
  fail the baked-var guard) with the deliberately non-baked INFLEXA_DEV=1
  runtime escape hatch; truth table unit-tested and verified live at the
  source runtime (release surface = the product commands exactly)
- adversarial verify (FIX-FIRST on doc drift, then clean): CLAUDE.md/
  CONTEXT.md no longer teach the retired module or the session-scoped event
  contract; spec wording matches the real unregistered-command refusal;
  the chat-command requirement retitled to its standing dev-channel role
- findings: release binary builds are broken at HEAD by a PRE-EXISTING
  DBOS-SDK telemetry bundling failure (reproduced on unmodified HEAD) —
  tracked as a follow-up with the also-pre-existing INFLEXA_GIT_COMMIT
  bake gap

578 cli tests pass (0 fail); typecheck + eslint clean.
