+++ cli/.env.example
@@ -1,3 +1,7 @@
 INFLEXA_AUTH0_DOMAIN=dev-elbhhfoh0p2nshbt.us.auth0.com
 INFLEXA_AUTH0_CLIENT_ID=ou9qEV6dw74rksE2UxosbJ0STQrqaGR2
 INFLEXA_AUTH0_AUDIENCE=bpOwTfm8m69QsWD-p-SYAJFbzP-loQg_dZDAehLeiTNOELwAetR7VprsAlXQM-id
+# Baked build channel. Leave unset for `bun run dev` (dev channel: chat/profile/run registered).
+# Release builds MUST set INFLEXA_BUILD_CHANNEL=release so those dev commands are omitted from the
+# binary; the build's missing-var guard rejects an empty/unset value on a release build.
+# INFLEXA_BUILD_CHANNEL=release
