+++ cli/CLAUDE.md
@@ -184,7 +184,7 @@ Code is grouped **by feature (vertical slice)**, not by technical layer: a featu
 
 - `src/index.ts` — entry point: telemetry/log wiring, shutdown hooks, then `cli.parse()`
 - `src/cli/` — the cac command **registry** (`index.ts`) + help formatting, nothing else. Each command lazy-imports its action: text commands from their module (e.g. `import("../modules/auth/login.ts")`), TUI screens from `tui/` (e.g. `import("../tui/app.launch.tsx")`).
-- `src/modules/<domain>/` — feature slices (see [Modules](#modules)): **headless** domain logic + the text command actions that drive them. Interactive views are NOT here (see `tui/`). Today: `auth/` (Auth0 device flow + `login`/`logout`/`whoami`), `proxy/` (CLIProxyAPI lifecycle + `setup`), `intelligence/` (model-interaction chat engine + `sessions` command), `anchor/` (folder-identity markers + lazy path reconciliation).
+- `src/modules/<domain>/` — feature slices (see [Modules](#modules)): **headless** domain logic + the text command actions that drive them. Interactive views are NOT here (see `tui/`). Today: `auth/` (Auth0 device flow + `login`/`logout`/`whoami`), `proxy/` (CLIProxyAPI lifecycle + `setup` + the proxy model helpers in `models.ts`), `analysis/` (analysis lifecycle + the `sessions` command), `anchor/` (folder-identity markers + lazy path reconciliation).
 - `src/db/` — shared SQLite layer: `primary.ts` (connection), `primary_migrations.ts`, `primary_query.ts`, `primary_mutation.ts`, `errors.ts`, `util.ts`. Queries/mutations stay here (verb-split, beside the migrations); a module imports the functions it needs.
 - `src/tui/` — the **presentation layer / app shell**: the entry app plus shared, app-level, or reusable Solid/opentui code. Today: `app.tsx` (the root chat screen, launched by every command that opens a chat), `app.launch.tsx` (`launchTui`), `command_palette.tsx` / `commands.tsx` (the palette adapter + command registry), `app_config.tsx` (settings — an app-level screen), `theme.ts` (the reactive accessor — the id list + palette data live in `lib/design_system.ts`; also home to the shared `Notice` type + `noticeColor` mapping, since a notice kind maps onto a palette role), and `components/` (shared, domain-agnostic widgets — see below). Presentation sits *above* the logic modules: it may import module logic (view → logic); modules must never import `tui/`. **Where a view lives** mirrors Lumen's `components/` vs `modules/<m>/components/` split — shared / app-shell / app-level screens go here; a view owned by exactly one feature co-locates in that module. `app_config.tsx` is an app-level exception that lives here — not a license to put every view in `tui/`. Add `tui/<domain>/` (or module-side view folders) when a screen outgrows one file.
   - `src/tui/components/` — shared TUI widgets. A widget belongs here **iff** it (a) imports only `theme` + opentui/solid (no `modules/`, `db/`, or other domain imports) and (b) has ≥2 callers; one component per file, no barrels. A feature-coupled adapter (e.g. `CommandPalette`, which maps `Command` objects) stays in the `tui/` app-shell, not here. Today: `dialog/` (the dialog subsystem — see below), `list_core.tsx` + `fixed_list.tsx` + `dynamic_list.tsx` (the list primitives — pure fuzzy-filtered grouped lists; `FixedList` reads its items ONCE and renders with `<For>`, `DynamicList` tracks reactive items and renders with `<Index>`; hosts own the filter input and pass `query` down), `text_area.tsx` (`TextArea` — themed textarea with chrome tiers and mode tracking), `text_input.tsx` (`TextInput` — themed single-line input with per-keystroke callback). Both editors take an `autoFocus` opt-out (list-first hosts, gallery exhibits, showcased dialogs) that also seeds their chrome's focus signal, so a blurred mount renders blurred from the first frame.
@@ -200,7 +200,7 @@ Code is grouped **by feature (vertical slice)**, not by technical layer: a featu
 A module under `src/modules/<domain>/` groups everything about one domain: its logic, its CLI command action(s), and its logic-local types. There is **no mandated file layout** — add files as the domain needs them, not preemptively (named exports only, no barrels, per [Naming conventions](#naming-conventions)).
 
 - **Public surface.** Whatever other layers import is the module's API; import it directly from the owning file (e.g. `ensureProxyReady` from `modules/proxy/setup.ts`). No barrel/index re-exports.
-- **Dependency direction.** Modules import shared infra (`lib/`, `db/`, `src/types/`, `extensions/`) and **other modules, acyclically** (e.g. `intelligence` → `proxy`; a future `analysis` → `anchor`). They must **not** import `tui/` — presentation depends on logic, never the reverse. Infrastructure (`lib/`, `db/`) must **never** import a module. If two modules need the same code, lift it to a shared layer.
+- **Dependency direction.** Modules import shared infra (`lib/`, `db/`, `src/types/`, `extensions/`) and **other modules, acyclically** (e.g. `harness` → `proxy`; `analysis` → `anchor`). They must **not** import `tui/` — presentation depends on logic, never the reverse. Infrastructure (`lib/`, `db/`) must **never** import a module. If two modules need the same code, lift it to a shared layer.
 
 ## Global extensions
 
@@ -216,7 +216,7 @@ Reach for an extension only when a helper is genuinely cross-cutting; keep singl
 There is **one bus** (`Bus` in `src/lib/bus.ts`), shared by every domain — session, provenance, and any future concern. Domain separation is by event `type` string, not by bus instance. **Never add a second bus** to separate concerns; if the current event contract feels wrong, the fix is better types, not more buses.
 
 - **One event type per domain action.** Each `BusEvent` member carries exactly the fields its action needs. Never pack multiple sub-actions into a single event discriminated by an interior field with nullable companions — that defeats TypeScript's narrowing and forces consumers to guard against impossible states. E.g. provenance has `prov.analysis_created`, `prov.input_added`, `prov.input_removed` — not one `prov.recorded` with a nullable `input`.
-- **Event types live in `src/types/events.ts`** — the `BusEvent` discriminated union is the contract. Session-scoped events carry `sessionId`; analysis-scoped events carry `analysisId`. Consumers filter by `type`.
+- **Event types live in `src/types/events.ts`** — the `BusEvent` discriminated union is the contract. Every member today is analysis-scoped provenance (`prov.*`, carrying `analysisId`); the harness conversation path writes the Solid store directly and does not use the bus. Consumers filter by `type`.
 - **Design rationale:** a dedicated bus per domain only earns its keep when that domain needs its own subscriber lifecycle, backpressure, or error isolation. inflexa's bus is a fire-and-forget notification channel with one subscriber per concern — multiplying buses adds wiring overhead for no structural gain. (Validated against OpenCode, which routes ~80+ event types across all domains through a single typed bus.)
 
 ## Solid + opentui (TUI)
@@ -264,7 +264,7 @@ The remedy: **any fixed chrome row placed directly below a `flexGrow` scrollbox 
 The bus contract and design rationale live in [Event bus — single bus, typed events](#event-bus--single-bus-typed-events) above. TUI-specific rules:
 
 - Subscribe in component setup with `Bus.on("inflexa", handler)` and always pair with `onCleanup(() => Bus.off("inflexa", handler))`.
-- Handlers must filter events by `sessionId` before applying them (provenance events carry `analysisId` and should be ignored).
+- Handlers must filter events by `analysisId` (every bus member is analysis-scoped provenance) and apply only the domains they own.
 
 ### Colors
