+++ cli/openspec/changes/archive/2026-07-08-retire-proxy-chat-dev-umbrella/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-07-08
