+++ cli/src/modules/analysis/sessions.ts
@@ -1,7 +1,13 @@
 import * as query from "../../db/primary_query.ts";
 import { dieOn } from "../../lib/cli.ts";
 
-export async function listSessions() {
+/**
+ * `inflexa sessions` — print the saved chat sessions (id, title, creation time).
+ * Sessions are the live launch-identity rows (a chat thread binds 1:1 to one), so this
+ * lists current data; their `messages`/`parts` history is frozen legacy data with no
+ * remaining writer. The command performs reads only — no row is created or modified.
+ */
+export async function listSessions(): Promise<void> {
     query.listSessions().match((sessions) => {
         if (sessions.length === 0) {
             console.log("No sessions found.");
